cleanup: isolate trusted catalogue snapshot ingestion
